Download docker-compose.yml if it doesn't exist
It doesn't implement "download" command and doesn't handle update of the
config file. It only checks if the config exists and if it doesn't
downloads it from master branch. So the cli can be usable. Currently if
a one downloads it and try to run there is an error:

```
ERROR:
        Can't find a suitable configuration file in this directory or any
        parent. Are you in the right directory?

        Supported filenames: docker-compose.yml, docker-compose.yaml

exit status 1
```

Signed-off-by: Maxim Sukharev <max@smacker.ru>