Create an empty default docker-compose.override.yml
When initializing a workdir, it will be linked the default
docker-compose.override.yml file, that could be used to override
sourced config.

If the default docker-compose.override.yml is deleted, it will
be restored everytime an init command is run, same as it is
currently done with the main docker-compose.yml.

Signed-off-by: David Pordomingo <David.Pordomingo.F@gmail.com>