Disallow re-initialization with differen value of --with-forks
- Switching --with-forks flag for already initialized working directory
is unsupported because gitcollector/ghsync can't delete data. Also user
might have some pre-computed data that would cause charts to fail if
forks are removed. The only safe way to switch the flag is to prune and
recreate. Automatic prune might be very unexpected for a user so it
shows an error and asks the user to run prune manually.

- Implement envFile decoder, it will be useful for showing status later
as well

Example of the error when an org was already initialized without forks:
```
$ go run cmd/sourced/main.go init orgs golang-migrate --with-forks
initialization failed: workdir was previously initialized with a different value for forks support
Cannot perform this action, full re-initialization is needed, run 'prune' command first
exit status 1
```

Signed-off-by: Maxim Sukharev <max@smacker.ru>